+++ m_swap.h
@@ -33,9 +33,9 @@
 // WAD files are stored little endian.
 #ifdef __BIG_ENDIAN__
 short	SwapSHORT(short);
-long	SwapLONG(long);
+long long	SwapLONG(long long);
 #define SHORT(x)	((short)SwapSHORT((unsigned short) (x)))
-#define LONG(x)         ((long)SwapLONG((unsigned long) (x)))
+#define LONG(x)         ((long long)SwapLONG((unsigned long long) (x)))
 #else
 #define SHORT(x)	(x)
 #define LONG(x)         (x)
